Move the creation of the test organisation to the pull controller
