+++ app/src/main/java/org/eyeseetea/malariacare/data/sync/importer/PullController.java
@@ -23,8 +23,10 @@ import android.content.Context;
 import android.os.AsyncTask;
 import android.util.Log;
 
+import org.eyeseetea.malariacare.R;
 import org.eyeseetea.malariacare.data.IDataSourceCallback;
 import org.eyeseetea.malariacare.data.database.model.OrgUnit;
+import org.eyeseetea.malariacare.data.database.model.Organisation;
 import org.eyeseetea.malariacare.data.database.utils.populatedb.PopulateDB;
 import org.eyeseetea.malariacare.data.remote.PullDhisSDKDataSource;
 import org.eyeseetea.malariacare.data.remote.SdkQueries;
@@ -102,6 +104,7 @@ public class PullController implements IPullController {
 
         if (isDemo) {
             createDummyOrgUnitsDataInDB();
+            addTestOrganisation();
         }
     }
 
@@ -116,6 +119,13 @@ public class PullController implements IPullController {
                 e.printStackTrace();
             }
         }
+    }
+
+    private void addTestOrganisation() {
+        Organisation testOrganisation = new Organisation();
+        testOrganisation.setName(mContext.getString(R.string.test_organisation_name));
+        testOrganisation.setUid(mContext.getString(R.string.test_organisation_uid));
+        testOrganisation.insert();
     }
 
     private void pullData(PullFilters pullFilters, List<OrganisationUnit> organisationUnits,
